requirements.txt
streamlit==0.71.0
pandas==1.1.4
torchvision==0.8.1
numpy==1.19.4
torch==1.7.0
opencv_python==4.4.0.46
Pillow==8.0.1
